Fix type mismatch in ShaderStage::source method
Changed the size parameter in glShaderBinary to int32_t to ensure compatibility with the function's expected parameter type, preventing potential runtime errors or warnings. This ensures the shader binary is correctly handled based on its size.

+++ code/graphics/opengl/shader_stage.cpp
@@ -21,7 +21,7 @@ namespace engine::gl
 
     void ShaderStage::source(const std::vector<char>& source) const
     {
-        glShaderBinary(1, &_handle, shader_binary_format, source.data(), source.size());
+        glShaderBinary(1, &_handle, shader_binary_format, source.data(), static_cast<int32_t>(source.size()));
         glSpecializeShader(_handle, "main", 0, nullptr, nullptr);
     }
 }
